refactor(plugin-openclaw): extract the delegate namespace binding store
delegate-runtime.ts crossed the 1200 LOC cap for non-grandfathered files;
the per-session binding store and its legacy migration are a storage
concern of their own.

Part of #3073
